% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterpie.R
\name{geom_scatterpie}
\alias{geom_scatterpie}
\title{geom_scatterpie}
\usage{
geom_scatterpie(mapping, data, cols, scale_fun = function(x) x/sum(x), ...)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{cols}{cols the pie data}

\item{scale_fun}{scale function for pie size}

\item{...}{additional parameters}
}
\value{
layer
}
\description{
scatter pie plot
}
\author{
guangchuang yu
}

