% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_scatterpie_legend.R
\name{geom_scatterpie_legend}
\alias{geom_scatterpie_legend}
\title{geom_scatterpie_legend}
\usage{
geom_scatterpie_legend(
  radius,
  x,
  y,
  n = 5,
  labeller,
  label_position = "right",
  ...
)
}
\arguments{
\item{radius}{radius vector}

\item{x}{x position}

\item{y}{y position}

\item{n}{number of circle}

\item{labeller}{function to label radius}

\item{label_position}{a character string indicating the position of labels,
"right" (default) or "left" or any abbreviation of these}

\item{...}{other text arguments passed on to \code{\link[ggplot2:layer]{ggplot2::layer()}}}
}
\value{
layer
}
\description{
legend of scatterpie
}
\author{
Guangchuang Yu
}
