% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recenter.R
\name{recenter}
\alias{recenter}
\title{recenter}
\usage{
recenter(mapdata, center, longitude_column = "long", clean = FALSE)
}
\arguments{
\item{mapdata}{map data, shoud be a data.frame}

\item{center}{center}

\item{longitude_column}{longitude column}

\item{clean}{whether clean data to prevent cross lines}
}
\value{
updated map data
}
\description{
re-center map data
}
\author{
ygc
}

