% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterpie-package.R
\docType{package}
\name{scatterpie-package}
\alias{scatterpie}
\alias{scatterpie-package}
\title{scatterpie: Scatter Pie Plot}
\description{
Creates scatterpie plots, especially useful for plotting pies on a map.
}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID})

Other contributors:
\itemize{
  \item Shuangbin Xu \email{xshuangbin@163.com} (\href{https://orcid.org/0000-0003-3513-5362}{ORCID}) [contributor]
}

}
\keyword{internal}
